requirements.txt
ads
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alembic
annotated-types
anyio
archspec
argcomplete
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asdf
asdf-astropy
asdf_coordinates_schemas
asdf_standard
asdf_transform_schemas
asdf_wcs_schemas
assist
asteval
astroplan==0.10.1
astropy
astropy-iers-data
astroquery==0.4.7
asttokens
async-lru==2.0.4
attrs
autograd
black
pre-commit
aws-xray-sdk
babel==2.16.0
backoff
backports.tarfile
batoid
batoid_rubin
bcrypt
beautifulsoup4
bleach==6.2.0
blinker
boltons
boto3
botocore
Bottleneck
Brotli
cached-property
cachetools
cassandra-driver==3.29.2
certifi
cffi
charset-normalizer
cli_helpers
click
colorama
coloredlogs
colour-science
comm
configobj
configparser
confluent-kafka
contourpy
corner==2.2.3
coverage
coveralls
cryptography
cycler
danish
debugpy==1.8.11
decorator
deepdiff
defusedxml
Deprecated
dill
distro
docopt==0.6.2
dogpile.cache
dustmaps
emcee
esutil
exceptiongroup
execnet
executing
fastavro
filelock
firefly_client
fitsio
flake8
Flask
Flask-Cors
fonttools
fqdn==1.5.1
frozendict
frozenlist
fsspec
future
GalSim
geomet
getCalspec
ginga
gitdb==4.0.12
GitPython==3.1.44
globus-sdk
gmpy2
greenlet
gwcs
h11
h2
h5py
healpy==1.17.3
healsparse
hpack==4.0.0
hpgeom
html5lib
httpcore
httpx
huggingface_hub
humanfriendly
humanize
hyperframe
hypothesis
idds-client
idds-common
idds-doma
idds-workflow
idna
imagecodecs
imageio
iminuit
importlib_metadata
importlib_resources
iniconfig
ipykernel==6.29.5
ipython
ipywidgets
isoduration==20.11.0
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
jedi
jeepney
Jinja2
jmespath
joblib
joserfc
json5==0.10.0
jsondiff
jsonpatch
jsonpointer
jsonschema
jsonschema-path
jsonschema-specifications
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets
keyring
kiwisolver
lazy-object-proxy
lazy_loader
lightning-utilities==0.11.9
llvmlite==0.43.0
lmfit
LSSTDESC.Coord
lxml
Mako
MarkupSafe
matplotlib==3.9.2
matplotlib-inline
mpi4py
mpmath
multidict
multiprocess
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest_asyncio
networkx
ngmix
nose
notebook==7.3.2
notebook_shim==0.2.4
numba
numexpr
numpy
openapi-schema-validator
openapi-spec-validator
opencv-python
opencv-python-headless
opt_einsum
orderly-set
packaging
panda-client
pandas
pgspecial
photutils
pickleshare
Piff
pillow
pyarrow==17.0.0
pyparsing
pyproj
PySocks
pysynphot
pytest
pytest-arraydiff
pytest-astropy
pytest-astropy-header
pytest-cov
pytest-doctestplus
pytest-filter-subpackage
pytest-mock
pytest-openfiles
pytest-remotedata
pytest-runner
pytest-session2file
pytest-subtests
pytest-vcr
pytest-xdist
python-dateutil
python-git==2018.2.1
python-json-logger==3.2.1
python-utils
pytorch-lightning==2.5.0.post0
pytz
pyvo
PyWavelets==1.7.0
PyYAML
pyzmq
QtPy
rebound
referencing
requests
responses
rfc3339-validator
rfc3986-validator==0.1.1
rpds-py
ruamel.yaml
ruamel.yaml.clib
rucio-clients
ruff
s3fs
s3transfer
sbpy
schwimmbad
scikit-image
scikit-learn
scipy
SCons
seaborn
SecretStorage
semantic-version
Send2Trash==1.8.3
setproctitle
setuptools-scm
six
skyproj
smatch
smmap==5.0.2
sniffio
sorcha
sortedcontainers
soupsieve
spiceypy
SQLAlchemy
sqlparse
stack-data
statsmodels
stevedore
stomp.py
sympy==1.13.1
synphot
tables
tabulate
tblib
terminado==0.18.1
threadpoolctl
tifffile
timm
time-machine
tinycss2==1.4.0
toml
tomli
tornado
tqdm
traitlets
TreeCorr
treegp
trimesh
triton
truststore
ts_xml
typeguard
types-python-dateutil==2.9.0.20241206
types-PyYAML
typing_extensions
tzdata
uncertainties
uri-template==1.3.0
urllib3
vcrpy
wcwidth
webcolors==24.11.1
webdav4
webencodings
websocket-client
Werkzeug
wget==3.2
widgetsnbextension
wrapt
ws4py==0.5.1
xmltodict
yarl
zenodo_get
zipp
zstandard==0.23.0
